==> study/ClustersEnsemble/ExcITPTSubtypeEnsemble/2.related_dmrs.ipynb <==
import pandas as pd
import anndata

## Parameters

use_clusters = [
    'CA1_Ak5', 'CA1_Chrm3', 'CA1_Kif26a', 'CA1_Ptprg',
    'CA3_Cadm2', 'CA3_Efnb2', 'CA3-St18_Epha5', 'CA3-St18_Nuak1',
    'CA3-St18_Tead1', 'DG_dg-all', 'DG-po_Bcl11a', 'DG-po_Calb2',
    'DG-po_Kctd8', 'Gfra1_Gfra1', 'IG-CA2_Chrm3', 'IG-CA2_Peak1', 'IG-CA2_Xpr1'
]
hypo_dmr_hits_path = '/home/hanliu/project/mouse_rostral_brain/DMR/SubType/Total/TotalHits.HypoDMR.h5ad'
# 'CA1_Lingo2' not include, because it co-cluster with RHP, probably an incomplete cluster

# Parameters
use_clusters = ["CLA_Bcl11a", "CLA_Cdh8", "CLA_Nrp2", "EP_Adcy8", "EP_Rgs8", "EP_Tspan5", "IT-L23_Cux1", "IT-L23_Foxp1", "IT-L23_Ptprt", "IT-L23_Tenm2", "IT-L4_Astn2", "IT-L4_Shc3", "IT-L5_Cdh8", "IT-L5_Etv1", "IT-L5_Grik3", "IT-L6_Cadps2", "IT-L6_Fstl4", "IT-L6_Man1c1", "IT-L6_Oxr1", "PT-L5_Abca12", "PT-L5_Astn2", "PT-L5_Kcnh1", "PT-L5_Nectin1", "PT-L5_Plcb4", "PT-L5_Ptprt", "PT-L5_Tenm2", "PT-L5_Tmtc2", "PT-L5_Unc5b"]
hypo_dmr_hits_path = "/home/hanliu/project/mouse_rostral_brain/DMR/SubType/Total/TotalHits.HypoDMR.h5ad"


## Hypo Hits

hypo_hits_adata = anndata.read_h5ad(hypo_dmr_hits_path)
related_dmr = (hypo_hits_adata[:, use_clusters].X.sum(axis=1) > 0).A1
use_dmr = hypo_hits_adata.obs_names[related_dmr]

use_hits = hypo_hits_adata[use_dmr][:, use_clusters]
use_hits.write_h5ad('RelatedClusterAndDMR.hypo_hits.h5ad')

use_hits

## DMR Rate

with pd.HDFStore('/home/hanliu/project/mouse_rostral_brain/DMR/SubType/Total/DMRInfo.h5') as hdf:
    dmr_rate = hdf['Rate']
dmr_rate.index = hypo_hits_adata.obs_names
use_rate = dmr_rate.loc[use_dmr, use_clusters]

use_rate.to_msgpack('RelatedClusterAndDMR.mcg_rate.msg')

with open('related_dmrs.txt', 'w') as f:
    for dmr in use_dmr:
        f.write(dmr + '\n')
