==> study/ClustersEnsemble/ExcITPTSubtypeEnsemble/3.related_corr.ipynb <==
import pandas as pd

## Parameter

corr_cutoff = 0.35
distance_cutoff = 500000

# Parameters
corr_cutoff = 0.3
distance_cutoff = 500000


total_corr = pd.read_msgpack('/home/hanliu/project/mouse_rostral_brain/study/DMRGeneCorr/TotalGeneDMRCorrLoop.0.3.msg')

use_gene = pd.read_csv('related_dmgs.txt', index_col=0, header=None).index
use_dmr = pd.read_csv('related_dmrs.txt', index_col=0, header=None).index

use_corr = total_corr[total_corr['DMR'].isin(use_dmr) & 
           total_corr['Gene'].isin(use_gene) & 
           (total_corr['DMR_to_gene_dist'].abs() < distance_cutoff) & 
           (total_corr['Corr'] > corr_cutoff)].reset_index(drop=True)

use_corr.to_msgpack('RelatedCorr.msg')

use_corr.shape

use_corr.head()
