==> study/ClustersEnsemble/ExcITPTSubtypeEnsemble/6.DMRGroupBasedOnGeneClustering.ipynb <==
import pandas as pd
import numpy as np
import anndata
import json
from collections import defaultdict

gene_clusters = anndata.read_h5ad('GeneClustering.h5ad')
gene_clusters = gene_clusters.obs['leiden']

total_corr = pd.read_msgpack('RelatedCorr.msg')

dmr_records = []
with pd.HDFStore('FinalDMGDMR.h5') as f:
    for k in f.keys():
        if k.endswith('Corr'):
            corr_index = f[k]
        else:
            continue
        this_corr = total_corr.loc[corr_index].copy()
        this_corr['gene_cluster'] = this_corr['Gene'].map(gene_clusters)
        dmr_records.append(this_corr[['DMR', 'gene_cluster']])
dmr_records = pd.concat(dmr_records).reset_index(drop=True).drop_duplicates()

cluster_dmr_index = dmr_records.groupby('gene_cluster').apply(lambda i: i['DMR'].tolist()).to_dict()

with open('GeneCluster.relatedDMR.index.json', 'w') as f:
    json.dump(cluster_dmr_index, f)

for cluster, dmrs in cluster_dmr_index.items():
    print(f'Cluster {cluster}', len(dmrs), sep='\t')
